[Added] Patch 7/17/2017

+++ MMOCoreORB/src/server/zone/objects/tangible/component/lightsaber/LightsaberCrystalComponentImplementation.cpp
@@ -75,7 +75,7 @@ void LightsaberCrystalComponentImplementation::fillAttributeList(AttributeListMe
 			alm->insertAttribute("wpn_attack_cost_health", sacHealth);
 			alm->insertAttribute("wpn_attack_cost_action", sacAction);
 			alm->insertAttribute("wpn_attack_cost_mind", sacMind);
-			alm->insertAttribute("forcecost", (int)getForceCost());
+			alm->insertAttribute("forcecost", (float)getForceCost());
 		} else {
 			StringBuffer str3;
 			str3 << "@jedi_spam:saber_color_" << getColor();
@@ -87,7 +87,7 @@ void LightsaberCrystalComponentImplementation::fillAttributeList(AttributeListMe
 			alm->insertAttribute("wpn_attack_cost_health", sacHealth);
 			alm->insertAttribute("wpn_attack_cost_action", sacAction);
 			alm->insertAttribute("wpn_attack_cost_mind", sacMind);
-			alm->insertAttribute("forcecost", (int)getForceCost());
+			alm->insertAttribute("forcecost", (float)getForceCost());
 		}
 	}
 
@@ -100,7 +100,7 @@ void LightsaberCrystalComponentImplementation::fillAttributeList(AttributeListMe
 			alm->insertAttribute("wpn_attack_cost_health", sacHealth);
 			alm->insertAttribute("wpn_attack_cost_action", sacAction);
 			alm->insertAttribute("wpn_attack_cost_mind", sacMind);
-			alm->insertAttribute("forcecost", (int)getForceCost());
+			alm->insertAttribute("forcecost", (float)getForceCost());
 		} else {
 			StringBuffer str;
 			str << "@jedi_spam:crystal_quality_" << getQuality();
